requirements.txt
pwntools
pillow